CORE-1360: Refactor fetchers
