+++ src/server/admin/teams/controllers/add/add-member.js
@@ -3,7 +3,7 @@ import qs from 'qs'
 import { buildErrorDetails } from '../../../../common/helpers/build-error-details.js'
 import { sessionNames } from '../../../../common/constants/session-names.js'
 import { addUserValidation } from '../../helpers/schema/add-user-validation.js'
-import { addMemberToTeam } from '../../helpers/fetch/index.js'
+import { addMemberToTeam } from '../../helpers/fetch/fetchers.js'
 import { pluralise } from '../../../../common/helpers/pluralise.js'
 
 const addMemberController = {
