+++ src/server/admin/teams/controllers/delete/confirm-delete-team.js
@@ -2,7 +2,7 @@ import Joi from 'joi'
 import Boom from '@hapi/boom'
 import { teamIdValidation } from '@defra/cdp-validation-kit'
 
-import { fetchCdpTeam } from '../../helpers/fetch/index.js'
+import { fetchCdpTeam } from '../../helpers/fetch/fetchers.js'
 import { transformTeamToSummary } from '../../transformers/team-to-summary.js'
 import { transformTeamUsersToTaskList } from '../../transformers/team-users-to-task-list.js'
