+++ src/server/admin/teams/controllers/team.js
@@ -1,7 +1,7 @@
 import Joi from 'joi'
 import Boom from '@hapi/boom'
 
-import { fetchCdpTeam } from '../helpers/fetch/index.js'
+import { fetchCdpTeam } from '../helpers/fetch/fetchers.js'
 import { transformTeamToSummary } from '../transformers/team-to-summary.js'
 import { transformTeamUsersToTaskList } from '../transformers/team-users-to-task-list.js'
 import { transformTeamScopesToTaskList } from '../transformers/team-scopes-to-task-list.js'
