--- src/server/admin/teams/helpers/fetch/edit-team.js
@@ -1,20 +1,0 @@
-import { config } from '../../../../../config/config.js'
-import { removeNil } from '../../../../common/helpers/remove-nil.js'
-
-function editTeam(request, teamId, payload) {
-  const endpoint = config.get('userServiceBackendUrl') + '/teams/' + teamId
-
-  return request.authedFetchJson(endpoint, {
-    method: 'patch',
-    payload: removeNil({
-      name: payload.name,
-      description: payload.description,
-      serviceCodes: payload.serviceCodes,
-      alertEmailAddresses: payload.alertEmailAddresses,
-      alertEnvironments: payload.alertEnvironments,
-      github: payload.github
-    })
-  })
-}
-
-export { editTeam }
