--- src/server/admin/teams/helpers/fetch/fetch-cdp-team.js
@@ -1,11 +1,0 @@
-import { config } from '../../../../../config/config.js'
-import { fetchJson } from '../../../../common/helpers/fetch/fetch-json.js'
-
-async function fetchCdpTeam(teamId) {
-  const endpoint = config.get('userServiceBackendUrl') + `/teams/${teamId}`
-
-  const { payload } = await fetchJson(endpoint)
-  return payload
-}
-
-export { fetchCdpTeam }
