--- src/server/admin/teams/helpers/fetch/index.js
@@ -1,17 +1,0 @@
-import { fetchCdpTeam } from './fetch-cdp-team.js'
-import { searchCdpUsers } from './search-cdp-users.js'
-import { addMemberToTeam } from './add-member-to-team.js'
-import { removeMemberFromTeam } from './remove-member-from-team.js'
-import { searchGithubTeams } from './search-github-teams.js'
-import { fetchCdpTeams } from './fetch-cdp-teams.js'
-import { editTeam } from './edit-team.js'
-
-export {
-  editTeam,
-  fetchCdpTeams,
-  searchGithubTeams,
-  fetchCdpTeam,
-  searchCdpUsers,
-  addMemberToTeam,
-  removeMemberFromTeam
-}
