--- src/server/admin/teams/helpers/fetch/remove-member-from-team.js
@@ -1,13 +1,0 @@
-import { config } from '../../../../../config/config.js'
-
-async function removeMemberFromTeam(request, teamId, userId) {
-  const endpoint =
-    config.get('userServiceBackendUrl') + `/teams/${teamId}/remove/${userId}`
-
-  const { payload } = await request.authedFetchJson(endpoint, {
-    method: 'patch'
-  })
-  return payload
-}
-
-export { removeMemberFromTeam }
