--- src/server/admin/users/helpers/fetch/fetch-cdp-user.js
@@ -1,11 +1,0 @@
-import { config } from '../../../../../config/config.js'
-import { fetchJson } from '../../../../common/helpers/fetch/fetch-json.js'
-
-async function fetchCdpUser(userId) {
-  const endpoint = config.get('userServiceBackendUrl') + `/users/${userId}`
-
-  const { payload } = await fetchJson(endpoint)
-  return payload
-}
-
-export { fetchCdpUser }
