--- src/server/admin/users/helpers/fetch/index.js
@@ -1,11 +1,0 @@
-import { fetchCdpUser } from './fetch-cdp-user.js'
-import { fetchCdpUsers } from './fetch-cdp-users.js'
-import { searchAzureActiveDirectoryUsers } from './search-azure-active-directory-users.js'
-import { searchGithubUsers } from './search-github-users.js'
-
-export {
-  fetchCdpUser,
-  fetchCdpUsers,
-  searchAzureActiveDirectoryUsers,
-  searchGithubUsers
-}
