--- src/server/admin/users/helpers/fetch/search-github-users.js
@@ -1,13 +1,0 @@
-import { config } from '../../../../../config/config.js'
-import { fetchJson } from '../../../../common/helpers/fetch/fetch-json.js'
-
-async function searchGithubUsers(query) {
-  const endpoint =
-    config.get('userServiceBackendUrl') +
-    `/github-users${query ? '?query=' + query : ''}`
-
-  const { payload } = await fetchJson(endpoint)
-  return payload
-}
-
-export { searchGithubUsers }
